fix: change the type to float32

+++ uniface/detection/scrfd.py
@@ -251,7 +251,7 @@ class SCRFD(BaseDetector):
 
         detections = pre_det[keep, :]
         landmarks = landmarks[order, :, :]
-        landmarks = landmarks[keep, :, :].astype(np.int32)
+        landmarks = landmarks[keep, :, :].astype(np.float32)
 
         if 0 < max_num < detections.shape[0]:
             # Calculate area of detections
